+++ shuup_tests/admin/test_media_module.py
@@ -7,7 +7,6 @@
 import json
 
 import pytest
-from django.contrib.auth.models import AnonymousUser
 from django.core.files.uploadedfile import InMemoryUploadedFile
 from django.http import JsonResponse
 from django.test.client import RequestFactory
@@ -22,24 +21,28 @@ from shuup_tests.utils import printable_gibberish
 
 
 @pytest.mark.django_db
-def test_media_view_images(rf):
+@pytest.mark.parametrize("is_public, expected_file_count", [(False, 0), (True, 1)])
+def test_media_view_images(rf, admin_user, is_public, expected_file_count):
     get_default_shop()
     folder = Folder.objects.create(name="Root")
-    file = File.objects.create(name="normalfile", folder=folder)
-    img = Image.objects.create(name="imagefile", folder=folder)
+    File.objects.create(name="normalfile", folder=folder)
+    img = Image.objects.create(name="imagefile", folder=folder, is_public=is_public)
 
     request = apply_request_middleware(rf.get("/", {"filter": "images", "action": "folder", "id": folder.id}))
+    request.user = admin_user
     view_func = MediaBrowserView.as_view()
     response = view_func(request)
     assert isinstance(response, JsonResponse)
     content = json.loads(response.content.decode("utf-8"))
-    assert len(content["folder"]["files"]) == 1
-    filedata = content["folder"]["files"][0]
-    assert filedata["name"] == img.name
+    assert len(content["folder"]["files"]) == expected_file_count
+    if expected_file_count:
+        filedata = content["folder"]["files"][0]
+        assert filedata["name"] == img.name
 
 
-def mbv_command(payload, method="post"):
+def mbv_command(user, payload, method="post"):
     request = RequestFactory().generic(method, "/")
+    request.user = user
     if method == "post":
         request._body = json.dumps(payload).encode("UTF-8")
     else:
@@ -48,11 +51,11 @@ def mbv_command(payload, method="post"):
     return json.loads(mbv(request).content.decode("UTF-8"))
 
 
-def mbv_upload(**extra_data):
+def mbv_upload(user, **extra_data):
     content = ("42" * 42).encode("UTF-8")
     imuf = InMemoryUploadedFile(BytesIO(content), "file", "424242.txt", "text/plain", len(content), "UTF-8")
     request = RequestFactory().post("/", dict({"action": "upload", "file": imuf}, **extra_data))
-    request.user = AnonymousUser()
+    request.user = user
     view = MediaBrowserView.as_view()
     response = view(request)
     return json.loads(response.content.decode("UTF-8"))
@@ -72,94 +75,94 @@ def get_id_tree(folders_response):
 
 
 @pytest.mark.django_db
-def test_new_folder():
+def test_new_folder(admin_user):
     folder1 = Folder.objects.create(name=printable_gibberish())
-    child_folder_data = mbv_command({"action": "new_folder", "name": "y", "parent": folder1.id})["folder"]
+    child_folder_data = mbv_command(admin_user, {"action": "new_folder", "name": "y", "parent": folder1.id})["folder"]
     assert Folder.objects.get(pk=child_folder_data["id"]).parent == folder1
-    root_folder_data = mbv_command({"action": "new_folder", "name": "y"})["folder"]
+    root_folder_data = mbv_command(admin_user, {"action": "new_folder", "name": "y"})["folder"]
     assert not Folder.objects.get(pk=root_folder_data["id"]).parent_id
 
 
 @pytest.mark.django_db
-def test_get_folder():
+def test_get_folder(admin_user):
     folder1 = Folder.objects.create(name=printable_gibberish())
     folder2 = Folder.objects.create(name=printable_gibberish(), parent=folder1)
-    root_resp = mbv_command({"action": "folder"}, "GET")
+    root_resp = mbv_command(admin_user, {"action": "folder"}, "GET")
     assert any(f["id"] == folder1.pk for f in root_resp["folder"]["folders"])
-    f1_resp = mbv_command({"action": "folder", "id": folder1.pk}, "GET")
+    f1_resp = mbv_command(admin_user, {"action": "folder", "id": folder1.pk}, "GET")
     assert f1_resp["folder"]["folders"][0]["id"] == folder2.pk
     assert not f1_resp["folder"]["files"]
     assert f1_resp["folder"]["name"] == folder1.name
-    assert mbv_command({"action": "folder", "id": -1}, "GET")["error"]
+    assert mbv_command(admin_user, {"action": "folder", "id": -1}, "GET")["error"]
 
 
 @pytest.mark.django_db
-def test_rename_folder():
+def test_rename_folder(admin_user):
     folder = Folder.objects.create(name=printable_gibberish())
-    mbv_command({"action": "rename_folder", "id": folder.pk, "name": "Space"})
+    mbv_command(admin_user, {"action": "rename_folder", "id": folder.pk, "name": "Space"})
     assert Folder.objects.get(pk=folder.pk).name == "Space"
 
 
 @pytest.mark.django_db
-def test_delete_folder():
+def test_delete_folder(admin_user):
     folder = Folder.objects.create(name=printable_gibberish())
-    mbv_command({"action": "delete_folder", "id": folder.pk})
+    mbv_command(admin_user, {"action": "delete_folder", "id": folder.pk})
     assert not Folder.objects.filter(pk=folder.pk).exists()
 
 
 @pytest.mark.django_db
-def test_rename_file():
+def test_rename_file(admin_user):
     file = File.objects.create(original_filename="test.jpg")
     assert force_text(file) == "test.jpg"
     assert file.pk
-    mbv_command({"action": "rename_file", "id": file.pk, "name": "test.tiff"})
+    mbv_command(admin_user, {"action": "rename_file", "id": file.pk, "name": "test.tiff"})
     file = File.objects.get(pk=file.pk)
     assert force_text(file) == "test.tiff"
 
 
 @pytest.mark.django_db
-def test_delete_file():
+def test_delete_file(admin_user):
     file = File.objects.create()
     assert file.pk
-    mbv_command({"action": "delete_file", "id": file.pk})
+    mbv_command(admin_user, {"action": "delete_file", "id": file.pk})
     assert not File.objects.filter(pk=file.pk).exists()
 
 
 @pytest.mark.django_db
-def test_move_file():
+def test_move_file(admin_user):
     folder1 = Folder.objects.create(name=printable_gibberish())
     folder2 = Folder.objects.create(name=printable_gibberish())
     file = File.objects.create(folder=folder1)
-    mbv_command({"action": "move_file", "file_id": file.pk, "folder_id": folder2.pk})
+    mbv_command(admin_user, {"action": "move_file", "file_id": file.pk, "folder_id": folder2.pk})
     assert File.objects.get(pk=file.pk).folder == folder2
-    mbv_command({"action": "move_file", "file_id": file.pk, "folder_id": 0})
+    mbv_command(admin_user, {"action": "move_file", "file_id": file.pk, "folder_id": 0})
     assert File.objects.get(pk=file.pk).folder is None
 
 
 @pytest.mark.django_db
-def test_auto_error_handling():
-    assert mbv_command({"action": "rename_file", "id": -1})["error"]
+def test_auto_error_handling(admin_user):
+    assert mbv_command(admin_user, {"action": "rename_file", "id": -1})["error"]
 
 
 @pytest.mark.django_db
-def test_upload(rf):
-    response = mbv_upload()
+def test_upload(rf, admin_user):
+    response = mbv_upload(admin_user)
     assert not File.objects.get(pk=response["file"]["id"]).folder
     folder = Folder.objects.create(name=printable_gibberish())
-    response = mbv_upload(folder_id=folder.pk)
+    response = mbv_upload(admin_user, folder_id=folder.pk)
     assert File.objects.get(pk=response["file"]["id"]).folder == folder
 
 
 @pytest.mark.django_db
-def test_upload_into_new_folder(rf):
+def test_upload_into_new_folder(rf, admin_user):
     assert Folder.objects.count() == 0
     # no folder
-    response = mbv_upload(path="/")
+    response = mbv_upload(admin_user, path="/")
     assert not File.objects.get(pk=response["file"]["id"]).folder
     assert Folder.objects.count() == 0
 
     # new folder
-    response = mbv_upload(path="/foo/bar")
+    response = mbv_upload(admin_user, path="/foo/bar")
     folder = File.objects.get(pk=response["file"]["id"]).folder
     assert folder.name == "bar"
     assert folder.parent.name == "foo"
@@ -167,7 +170,7 @@ def test_upload_into_new_folder(rf):
     assert Folder.objects.count() == 2
 
     # ensure path is correct with same folder names
-    response = mbv_upload(path="/bar/foo")
+    response = mbv_upload(admin_user, path="/bar/foo")
     folder = File.objects.get(pk=response["file"]["id"]).folder
     assert folder.name == "foo"
     assert folder.parent.name == "bar"
@@ -175,7 +178,7 @@ def test_upload_into_new_folder(rf):
     assert Folder.objects.count() == 4
 
     # upload into pre-existing folder
-    response = mbv_upload(path="/foo/bar")
+    response = mbv_upload(admin_user, path="/foo/bar")
     folder = File.objects.get(pk=response["file"]["id"]).folder
     assert folder.name == "bar"
     assert folder.parent.name == "foo"
@@ -183,7 +186,7 @@ def test_upload_into_new_folder(rf):
     assert Folder.objects.count() == 4
 
     # add subfolder
-    response = mbv_upload(path="/foo/bar/baz")
+    response = mbv_upload(admin_user, path="/foo/bar/baz")
     folder = File.objects.get(pk=response["file"]["id"]).folder
     assert folder.name == "baz"
     assert folder.parent.name == "bar"
@@ -192,7 +195,7 @@ def test_upload_into_new_folder(rf):
 
 
 @pytest.mark.django_db
-def test_get_folders(rf):
+def test_get_folders(rf, admin_user):
     # Create a structure and retrieve it
     folder1 = Folder.objects.create(name=printable_gibberish())
     folder2 = Folder.objects.create(name=printable_gibberish())
@@ -200,21 +203,21 @@ def test_get_folders(rf):
     folder4 = Folder.objects.create(name=printable_gibberish(), parent=folder2)
     folder5 = Folder.objects.create(name=printable_gibberish(), parent=folder3)
 
-    tree = get_id_tree(mbv_command({"action": "folders"}, "GET"))
+    tree = get_id_tree(mbv_command(admin_user, {"action": "folders"}, "GET"))
     assert set((folder1.id, folder2.id, folder3.id)) <= set(tree.keys())
     assert folder4.pk in tree[folder2.pk]
     assert folder5.pk in tree[folder3.pk]
 
 
 @pytest.mark.django_db
-def test_deleting_mid_folder(rf):
+def test_deleting_mid_folder(rf, admin_user):
     folder1 = Folder.objects.create(name=printable_gibberish())
     folder2 = Folder.objects.create(name=printable_gibberish(), parent=folder1)
     folder3 = Folder.objects.create(name=printable_gibberish(), parent=folder2)
-    tree = get_id_tree(mbv_command({"action": "folders"}, "GET"))
+    tree = get_id_tree(mbv_command(admin_user, {"action": "folders"}, "GET"))
     assert tree[folder1.pk] == {folder2.pk: {folder3.pk: {}}}
-    mbv_command({"action": "delete_folder", "id": folder2.pk})
-    tree = get_id_tree(mbv_command({"action": "folders"}, "GET"))
+    mbv_command(admin_user, {"action": "delete_folder", "id": folder2.pk})
+    tree = get_id_tree(mbv_command(admin_user, {"action": "folders"}, "GET"))
     assert tree[folder1.pk] == {folder3.pk: {}}
     folder1 = Folder.objects.get(pk=folder1.pk)
     assert list(folder1.get_children()) == [folder3]