+++ shuup_workbench/urls.py
@@ -7,10 +7,8 @@
 from django.conf import settings
 from django.conf.urls import include, url
 from django.conf.urls.static import static
-from django.contrib import admin
 
 urlpatterns = [
-    url(r'^admin/', include(admin.site.urls)),
     url(r'^sa/', include('shuup.admin.urls', namespace="shuup_admin", app_name="shuup_admin")),
     url(r'^api/', include('shuup.api.urls')),
     url(r'^', include('shuup.front.urls', namespace="shuup", app_name="shuup")),